requirements.txt
Flask>=1,<2
werkzeug>=1,<2
flask-session~=0.3.2

msal>=1.7,<2
azure-mgmt-resource==18.0.0
azure-mgmt-web==2.0.0
azure-cli-core
pywinrm>=0.3.0
wheel


ansible>=2.8.4
mysql
pyyaml
mysql-connector
pymysql
azure.cli
flask-mysqldb

pyodbc
adal
msrestazure
psycopg2
Flask-APScheduler

azure-mgmt-core<1.3.0,>=1.2.0
cryptography<3.4,>=3.2
pycrypto
requests==2.25.1